+++ firebase_rules/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Functions are now declared on `Match` statements instead of the annotation
 - `firestore.get<T>()` and `firestore.getAfter<T>()` now properly return a `FirestoreResource<T>`
 - The conversion to rules types is now `.rules()` instead of `.rules` to allow for casting iterables and maps
+- Enum conversion maps can now be passed into the `FirebaseRules` annotation
 
 ## 0.0.1+2
 - Readme fixes